+++ Sources/SourceKitLSP/SourceKitLSPServer.swift
@@ -917,7 +917,7 @@ extension SourceKitLSPServer {
       )
     )
     logger.log("Creating workspace at \(workspaceFolder.forLogging)")
-    logger.logFullObjectInMultipleLogMessages(header: "Options for workspace", options.loggingProxy)
+    logger.logFullObjectInMultipleLogMessages(header: "Workspace options", options.loggingProxy)
 
     let workspace = await Workspace(
       sourceKitLSPServer: self,
@@ -1003,7 +1003,7 @@ extension SourceKitLSPServer {
     }
 
     logger.log("Initialized SourceKit-LSP")
-    logger.logFullObjectInMultipleLogMessages(header: "SourceKit-LSP Options", options.loggingProxy)
+    logger.logFullObjectInMultipleLogMessages(header: "Global options", options.loggingProxy)
 
     await workspaceQueue.async { [hooks] in
       if let workspaceFolders = req.workspaceFolders {
@@ -2805,29 +2805,5 @@ fileprivate extension URL {
       return false
     }
     return other.pathComponents[0..<self.pathComponents.count] == self.pathComponents[...]
-  }
-}
-
-extension SourceKitLSPOptions {
-  /// We can't conform `SourceKitLSPOptions` to `CustomLogStringConvertible` because that would require a public import
-  /// of `SKLogging`. Instead, define an internal type that performs the logging of `SourceKitLSPOptions`.
-  struct LoggingProxy: CustomLogStringConvertible {
-    let options: SourceKitLSPOptions
-
-    var description: String {
-      options.prettyPrintedJSON
-    }
-
-    var redactedDescription: String {
-      options.prettyPrintedRedactedJSON
-    }
-  }
-
-  var loggingProxy: LoggingProxy {
-    LoggingProxy(options: self)
-  }
-
-  var forLogging: CustomLogStringConvertibleWrapper {
-    return self.loggingProxy.forLogging
   }
 }
